+++ climatedb/databases.py
@@ -1,24 +1,18 @@
 from collections import namedtuple, defaultdict
 from datetime import datetime
 from typing import List
-from typing import Optional
 
 from fastapi.templating import Jinja2Templates
 from pathlib import Path
-from pydantic import constr
 from rich import print
-from sqlalchemy import UniqueConstraint
-from sqlalchemy import select
 from sqlalchemy.sql.expression import func, select
-from sqlmodel import Field, Session, SQLModel, create_engine
+from sqlmodel import Session, SQLModel, create_engine
 import pandas as pd
 
-from climatedb import types
 from climatedb.config import data_home as home
 from climatedb.config import db_uri
 from climatedb.files import JSONLines
-from climatedb.types import Newspaper
-
+from climatedb.types import Article, AppTable, Newspaper
 
 
 def get_urls_for_paper(paper: str) -> List[str]:
@@ -65,33 +59,6 @@ def save_html(paper, article, response):
     )
     fi.parent.mkdir(exist_ok=True, parents=True)
     fi.with_suffix(".html").write_bytes(response.body)
-
-
-class Article(SQLModel, table=True):
-    __table_args__ = (UniqueConstraint("article_name"),)
-    id: Optional[int] = Field(default=None, primary_key=True)
-    body: constr(min_length=64)
-    headline: constr(min_length=8)
-    article_name: constr(min_length=4)
-    article_url: str
-    date_published: Optional[datetime]
-    date_uploaded: datetime
-    newspaper_id: int
-
-    article_length: int
-
-
-class AppTable(SQLModel, table=True):
-    #  article
-    body: constr(min_length=64)
-    headline: str
-    article_id: int = Field(primary_key=True)
-    article_url: str
-    date_published: Optional[datetime]
-    date_uploaded: datetime
-    #  newspaper
-    newspaper_id: int
-    fancy_name: str
 
 
 def find_id_for_newspaper(newspaper: str):
